refactor: Дополнительная декомпозиция main-page
